Created SmoothImageDiv general compnent, used on homepage hero.
The SmoothImageDiv component will make sure a div background image is fully loaded before fully displaying it.

The SmoothImageDiv component takes an image url, className, and a delay in milliseconds.

The component will apply the "loading" class to the div until the background image has been loaded. The transition and "loading" properties will need to be set in your style sheet.

It also allows a delay to be specified to ensure different SmoothImageDivs will fade in at the same time.

The images on the homepage will set opacity to 0 until the images are loaded, with a delay of 400 ms.